requirements.txt
jira==3.5.2
requests==2.31.0
pytz
google-api-python-client
google-auth-httplib2
google-auth-oauthlib